Extract function triggers to functions node in azure explorer
